Fix unsetting of copyright icon in FooterIcons
There's no $footericons['copyright']['copyright'] index in the
template data array and unset() will not raise any notice/warning.

The actual global, $wgFooterIcons in DefaultSettings.php has the index,
but this method is not working with its value directly, a similarly-named
method in SkinTemplate does. And it's what rearranges the array and
eventually change the original structure.

Bug: T266980
Change-Id: I8dc6babcdea9a8e20471a44cd025899a120b85a8