media: Filter out GPSAltitudeRef exif tag
The tag is processed to normalize the value of the GPSAltitude tag,
but sometimes sticks around in the output if the GPSAltitude is
missing or invalid.  The value of GPSAltitudeRef is single byte,
encoded as a string -- "\0" or "\1" -- but when accessing a file using
instant commons (or generally, using ForeignFileRepo) the "\0" is
munged by the action API which results in a bogus U+FFFD appearing in
the EXIF metadata output.

This patch reverts e6a04bf977ed9372e8b58a44765fde8edd1ec50b
(If85081f023a59b56557b9a74b8431c7be332d7e5) which addressed the same
issue (bogus GPSAltitudeRef values appearing on the File page) but did
it by 'supporting' the bogus U+FFFD value.  Better to preprocess the
EXIF metadata to remove the need for GPSAltitudeRef entirely.

Bug: T267370
Change-Id: Icb705c8f3248e14257532e2689731e67ead7c722