EditPage::runPostMergeFilters never starts with a hook error
EditPage::runPostMergeFilters used to run two hooks,
EditFilterMerged and EditFilterMergedContent, and checked
for a hook error after each. When the EditFilterMergedContent
hook was removed in 99800a9e0360e1a24c061d908dbe111bf687963f,
the check for a hook error created as a result of running that
hook was not removed.

Additionally, the hook error is already checked elsewhere
before the method is called.

Bug: T157658
Change-Id: I6704906c2abda2602f9994121f7cbd7c8a059eea