In SelectQueryBuilder add "joins" as an alias for join_conds
Add "joins" as an alias for join_conds in queryInfo(). Make the name
configurable in getQueryInfo().

It's actually a very common convention, used in CommentStore, LocalFile,
RevisionStore, etc.

join_conds is still needed, since it's used in ApiQueryBase, QueryPage
and IndexPager. Also it's the name of the IDatabase::select() parameter.

Change-Id: I4034ba70e9651d7ff12c72d7d75f3f9e78ffe67a