http: Remove exception from MultiHttpClient for missing cURL
This is both dead code and if it works, it contradicts the fact that
cURL library is not officially required by MediaWiki.

The function existence has already been checked in self::isCurlEnabled()
so it should be presumed to exist here because the call is guarded
behind that check.

The class level doc also said if cURL is not available, requests will be
made serially, it does not say Exception will be thrown which will
effectively makes cURL a hard dependency for MediaWiki and makes the
serial request code useless.

Follows-up 0584339f5ed (I2f9d4acbb49).

Change-Id: I7f172f177d7d17c80bb70b2fd6c67107f4340c21