Skin: Allow using link-class when class is an array
When creating a link with Skin::makeLink and specifying the
link-class option, makeLink blindly assumes the class attribute
it is given to be a string and will concat the value of link-class
to it, even when it is an array. PHP's type juggling will coerce
the array into a string and throw a notice.

Reproducable test case:

$link = [
	'text' => 'Link text',
	'href' => 'example.com',
	'class' => [ 'class1', 'class2' ]
];
$options = [
	'link-class' => 'link-class'
];

$skin->makeLink( 'test', $link, $options );

Change-Id: I6381e3e8893e4992dbed33c88674a5cf4367314f