Handle BlobAccessException exception
There are bad content blobs out there but not necessarily for all rows
here. If we fail to find a blob from a single address it should not
bring down the whole operation down. Instead, catch the exception and
print error msg then continue. The exception msg will advise the user
to use findBadBlobs.php script to remedy the issue.

Bug: T267468
Change-Id: I041d5c67bf241bb7cf2e31e0a4b0030e2cf9a13b