Add option to allow SpecialPageExecutor return the full HTML content
The property OutputPage::mBodytext does not always hold the full
HTML to be output only the body text which will be greatly altered
when Skin::outputPage is called.

I noticed this when a test that tries to check existence of a
'blocklink' in Special:Contributions subtitle consistently fails.

This happens because the links are only added when Skin::outputPage is
called but by calling OutputPage::getHTML() here and collecting the
buffer contents, the link is nowhere to be found since the skin method
will never be called if we didn't call OutputPage::output() (which -
will call the skin method)

Also OutputPage::output() already can return the fully generated html
via its optional parameter which we can just use here.

This patch adds optional parameter to specify that the caller wants the
entirety of the generated HTML. The default remains false, which
is to return OutputPage::getHTML().

Bug: T268322
Change-Id: I9b580e4ed93d989028d394803dba1ea7a9117592