rdbms: don’t teardown in LBFactorySingle destructor
This mirrors change I61914b9866 (commit 143333b172): since this class,
like LoadBalancerSingle, gets its connection injected rather than
managing its full lifetime, it should not close the connection when it
is destroyed. Without this change, the class cannot be used in
integration test, because MediaWikiIntegrationTest is not prepared for
its $this->db being closed.

Change-Id: I6cb62cfeef299b773cf62c16aa4d3d6f0496b7fe