SpecialEditWatchlist: remove unneeded BC hack
This hack was introduced with https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/core/+/c5a80bfb6600386b6fa248bf4903649fb367aed1%5E%21
in 2011, as SpecialEditWatchlist replaced WatchlistEditor, so that
callers could still use the old signature.

The class WatchlistEditor was removed in 1.24, and nobody should be
instantiating this class directly (outside of SpecialPageFactory), which
means that this hack is effectively unnecessary nowadays.

Change-Id: Iadf352b375b4caa7d88d01b15e2b8e73239ff924