Add new ArticleParserOptions hook
This hook allows discussion tools to toggle a parser option based on
artice title (namespace) and user identity.  In particular, it sets a
flag to generates reply links (a) on talk pages (b) if the user's
preferences opt in to discussion tools.  Other extensions may also
wish to be able to customize article parsing based on title or user.

Change-Id: I883a37fd67108243e7a20683b1a5d59fd0f6e39f