migrateComments: Cast array keys back to string before passing to the DB
My 2019 change to addQuotes() causes integers to be unquoted in SQL.
Putting a numeric string into an array key causes it to be implicitly
converted to integer. But comment text is properly a string and so
should be quoted. PostgreSQL gives an error if an integer is used for
a string column.

Bug: T268887
Change-Id: I28686ef7b3645b1fde3256d679945d7b48c13e7b