PageRestrictions: Remove unreachable logic
On another look after I242b1e7, I realized that this method will
not return null and the null-coalescing is effectively unreachable.
It will never be executed.

The property can indeed be null before the method is called, but
thereafter it can only be Title instance or false. It will only
resets to null after the object destruction.

Change-Id: I19b47affba49cded537011b6456d710bd55494b1