resourceloader: Remove redundant blob from validateScriptFile cache
Originally introduced in r91608 (0f201b19f47507), which cached
the input verbatim in Memcached, which is somewhat wasteful.
Instead, use null for the 99% case of the script being valid,
and cache that.

Also follows-up 3a748592f8da00 which made this a shared cache key,
which I don't think makes sense here since since there aren't any
shared resources going through this code path.

As counter example, the minification cache applies also to JS
files we deploy from Git, which are the same on all wikis and
identified by a file path that will match regardless of wiki
content, and so benefit from shared caching and can safely do so.

User scripts, however, are always stored on a particular wiki.
Even if they are loaded across domains in a browser, they will
still come through the wiki where they are stored, so this has
no added value.

Change-Id: I1615359e9ae8762f177004a02a9d3f69178e05c1