Selenium: fix selenium-daily NPM script
b8a2e21 removed hard-coded values for environment variables from
wdio.conf.js. That caused a couple of Jenkins jobs to fail because
MW_SCRIPT_PATH environment variable was not set.

This patch adds the variable to selenium-daily NPM script, executed by
the failing Jenkins jobs. The alternative is to add the variable to the
Jenkins job configuration.

Bug: T269755
Change-Id: I3c65db297b28449f000a8a79978ba777782f5ffa