Add skin versioning
As we further develop the capabilities of the skin, we will make
breaking changes as we fold functionality of extensions into
core (e.g. DismissableSiteNotice) and deprecate/change the data
provided in SkinMustache::getTemplateData

Providing a skin version which changes with major changes, allows
us to safely update extensions and skins.

It will be used immediately in I276325d

Change-Id: Ib4799e5380b472f6c2372c5982b24a404b756a7e