Set the default horizontal alignment based on the page content language
Previously it would flip based on the chrome, which isn't the way the
legacy parser has been doing it.

The choice of whether .mw-content-$dir are the right selectors can be
cleaned up with T258719.  We're already using them elsewhere in the file
so it seems like something that should be done all at once.

Bug: T269704
Change-Id: I1a2cc989c35506172af361cd86a4a378ddf58ecb