Prevent rcfilters menuselect widgets from winding up behind content on some skins
Set a z-index on the container so the unset z-index on the select
itself doesn't result in the select randomly winding up behind
bits of content, but also keeps it under the highlight selector.

Fixes CologneBlue, Example, and most third-party skins. Doesn't
appear to break skins that set their own custom z-index or that
otherwise didn't need this (Vector, Minerva, Modern, Timeless), but
someone may want to double check specific ones for oddities.

Conveniently also fixes the layering on Timeless so that the
widgets don't wind up above the fixed header, either.

Bug: T270351
Change-Id: Iecc6afab461ccbd096e8b05c652814d7c80be021