Selenium: Delete tests/selenium/selenium.sh
Also remove selenium NPM script that calls the file.

There was a time when both the file and the script were used.
As far as I know, both the file and the script are no longer used.
Selenium tests are now using selenium-test NPM script, both in
core and in extensions and skins.

Bug: T270419
Change-Id: I817f64b6d0a0efd64cd3fd9334e51336ceb9865f