[API] Ignore watchlist preferences for bot users
Every user matching User::isBot would be affected and not longer adding
pages to the watchlist along to its (default) preferences.
To override this behaviour the bot user must change watch explizit.

The watchlist is mostly not useful for bots doing massive edits or
uploads. This reduce the grow of the watchlist tables at all and avoids
less users with massive entries unable to delete or clear.

Bug: T258108
Change-Id: If76127315767bde70147197c88e93f51ca70edaa