DatabaseUpdater: Allow proper renaming of multiple indexes
With the abstract schema changes, indexes will now commonly be
renamed in group for the entire table (the single schema change).

Currently we are mis(using) DatabaseUpdater::renameIndex() which
was not meant for doing this and looks quite suboptimal.

In addition, because the renameIndex() method was purposefully
written to deal with a single index, the messages it prints
during the update become misleading and there's no reason for
to be so.

Change-Id: I04d93e13456cfbfbd46e922af3b6c4026cc36544