Bring back timestamp time of cl_timestamp instead of binary(14)
Fixing this is going to be complicated. We have to first turn this into
varbinary(20), write a maintenance script to fix the values, then make sure
mediawiki work properly with two value types at the same time and then
turning it to binary(14).

For now, let's unblock 1.36 release, we will follow up on fixing this
for good later.

Bug: T270032
Change-Id: Iabbe8f085812acd90536a06fee2ecd4db9b047d5