Improve REST API validation documentation
Make it clearer that the 'post' mode of ParamValidator only
applies to application/x-www-form-urlencoded or multipart/form-data
requests, and point to the equivalent mechanism for other request
types. Since developers coming from the action API are already
familiar with ParamValidator, this hopefully makes the other
validation mechanism easier to discover.

Change-Id: Iaa352a2d7e1d52a881b2cd960e34ebb0507ce83c