Update the autonym of language shi
This language is known in English as Shilha, Tashelhit,
Tachelhit, and also by some other names.

The most authoritative modern name in the language itself
is Taclḥiyt, with the letter c. It appears in the textbook
"Méthode de tachelhit, langue amazighe (berbère) du sud du Maroc"
by Abdallah El Mountassir and also in the online
"Dictionnaire Général de la Langue Amazighe Informatisé"
( https://tal.ircam.ma/dglai/lexieam.php ). It also directly
corresponds to the spelling in the Tifinagh script.

This change was requested by the localizers in translatewiki
and authors in the Wikipedia Incubator in this language.
They asked to define the Latin script as the primary.

The name of the variants are updated in the i18n file:
* The Latin names are changed to Taclḥit, as the Latin autonym.
* The Tifinagh name is added: it was missing, which caused
  the item in the dropdown to be invisible.

Change-Id: I23689a90460110570c54a05587e5084607268035