Prevent calling database method on null
It's allowed to call Database::getCacheSetOptions() with null
as a second argument since the paramater is nullable.

However, this will lead to fatal error as we will end up calling
getSessionLagStatus() on the passed null.

The only minor safeguard to this now is if the second argument is
omitted, then func_get_args() will not contain the default as it
returns only a copy of the passed arguments, but this is not
something that should be relied upon. There's nothing that can
prevent calling code from passing the second null.

Also pass the arguments directly to foreach() without
calling func_get_args() since we already know the possible
passable arguments and the function is not variadic.

Change-Id: I7d4804a715063d95827fcec6d14a3ae059b234c3