Restore use of func_get_args() in getCacheSetOptions()
The documentation says this can be used "for several DBs" but
because there are only two real paramaters and the gotcha  explained
at I7d4804a, I unintentionally limited it to work for "only two"
databases while fixing a different issue.

I am not sure wether there are callers calling with 5, 10 or 20
database objects, but PHP will never raise any warning even if
there are, the remaining args will just be silently discarded.

This restores the behavior so it can work "for several DBs", as
it used to.

Follow-up: I7d4804a715063d95827fcec6d14a3ae059b234c3
Change-Id: Id837e813cda19b86dcd14ec36b520f8465e5855d