Rewrite <langconvert> to support BCP 47 tags
This validates langconvert's "from" and "to" arguments as valid BCP 47 tags. For example, it will accept "sr-Cyrl" and "sr-cyrl" and reject the non-standard internal MediaWiki code "sr-ec". I made the BCP 47 matching case insensitive as that seems to conform with how MediaWiki handles it elsewhere and case sensitive matching would probably be a headache for users.

Bug: T271758
Change-Id: I9f765fe650279820d61c3a7e499ca99468df3d14