Use User::getEmail() not mEmail
The email address needs to be loaded from the CentralAuth hook, it's not
populated by User::load().

Fixed in isEmailConfirmed(), which is the subject of the bug, and also
two other places that looked relatively obvious and harmless. I'm not
fixing it in saveSettings() and addToDatabase() since the effect is not
so obvious. It's a bit of a can of worms.

Bug: T239523
Change-Id: I51b127ff92c6c3efe98c70e7590d43c7e7734568