Fix SpecialPage alias test
This was testing regex with completely unecessary long string.

I believe it was meant to be the test message instead of being the
haystack. Only the value of $alias may contain the slash, the rest
of the string has no any other slash, this is already known.

Also use simple string assertion instead of regex matching

Change-Id: I47eb0bfcd510e4f1c92099a9cb53b584a34036ce