WatchedItemStore: Use REPLACE when inserting expiries after page move
Sometimes when moving pages, a row for the target page may already exist
in the watchlist_expiry table. The exact scenario that's causing the
prod errors could not be repo'd in tests or locally, but using REPLACE
should be a safe solution.

Tests updated to actually verify the expiry of the new title, if
present, is overriden with the old one following the page move.

Bug: T271971
Change-Id: Id4d46fcdea4816e5714d266985ba86fdc6450cfb