rbdms: Improve signature of Database::getCacheSetOptions
One of the passed args is allowed to be null, but this null will
just be ignored and does not seems to be useful here so this is
now deprecated as suggested by Krinle at Id837e81

Now the function takes variadic args, which may still contain null
for sometime (owing to the method being public api).  In 1.37+, the
variadic parameter can be typehinted, and then the check for
instanceof would no longer be necessary.

Change-Id: I3863c3981470c1dbc11e417bc2252339bdee391f