Special:Contributions: Pass normalized target to ContribsPager
The point of instantiating ContribsPager previously had access to
$target variable which holds normalized ip-names, now that the pager
instantiation has moved to a different function for lazy-loading, we
need to pass the target to reach it.

We cannot use target option $this->opts['target'] because it contains
the non-normalized raw text from the request, and we cannot set the
prop after the normalization because getPager() may be called by
$this->contributionsSub() before then, and getPager() needs it.

There's one known caller of the protected $this->contributionsSub()
which will be fixed in a separate patch. SpecialContributions is
already not stable to be extended and this method is only accessible
via subclassing.

Bug: T272225
Change-Id: I0e4a1df75b05aecf32abe144664a93c87d1435b2