Add a temporary method for setting visibility of the block log entry
As per description on phabricator, BlockUser::log() creates a log entry
and inserts it, without offering callers any way to alter it. In particular,
it is not possible to alter visibility of the entry, which already has
at least a use case (T152394).

This method is @unstable because there haven't been enough public discussion
about it. It might then become @stable or be replaced by something else (e.g.
a constructor option or a hook), at the discretion of the maintainers
of this code (viz. AHT).

Bug: T271101
Change-Id: I4ed0c828da9c79de615fd8a88e99200fe9183fad