ImagePage: Compare thumbnail bounding box sizes to bounding box size
Currently, ImagePage::openShowImage compares the actual size of the main
thumbnail to the bounding boxes of the additional thumbnail links to
exclude duplicate links. However, rounding on portrait images can mean
that neither the height nor width of the thumbnail exactly match the
bounding box size.

This patch adds an additional comparison between the main thumbnail
bounding box and the additional size bounding boxes. The comparison
between the main thumbnail actual size and the additional size bounding
boxes remains to prevent any edge cases on that end. It may be
redundant.

Bug: T272201
Change-Id: Ib947f4a3ee26a1f885c0f38014fa23aa218e6ebe