UploadBase: Don't call MimeAnalyzer if mTempPath is null
As of PHP 8, fopen() throws ValueError if the path is null.
For now, keep the same behaviour as on PHP 7, which is that
MimeAnalyzer::doGuessMimeType sees fopen() return false (as
if the path was valid but doesn't exist or is not readable)
and falls back to unknown/unknown, which UploadBase ignored.

Bug: T272328
Change-Id: I5056eba2e3fd59d7971971a32b3de36f1a018603