RevisionStore: Prevent useless DB query
Ideally typehinting $id would be more effective, but a lot of
callers are just passing RevisionRecord::getId() directly which
is documented to return int|null. Typehinting the param here would
require first fixing those callers to ensure they are passing the
required int.

For now this patch will prevent the useless query, in case the
RevisionRecord::getId() call do return null where it's used as an
argument for this method. A case for <= 0 is also checked, while
not strictly necessary, it will also lead to useless query if such
value ever reaches here for whatever reason.

Bug: T270149
Change-Id: I5c23955e29efed4bec4c2e9652af1c2c3bf79218